requirements.txt
Django==3.1.7
djangorestframework==3.13.1
python-dotenv==0.15.0
requests==2.24.0
pytz